+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+* Include APM Trace saved views
+
 ## [v0.1.0]
 
 * Initial release of datadog Workflow
-
-## [Unreleased]
 
 [Unreleased]: https://github.com/lukewaite/alfred-datadog/compare/v0.1.0...HEAD
 [v0.1.0]: https://github.com/lukewaite/alfred-datadog/compare/v0.0.1...v0.1.0